feat(#226): add javadoc for FormattingDecompiler
