+++ src/main/java/org/eolang/opeo/DecompileMojo.java
@@ -32,7 +32,7 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.eolang.opeo.decompilation.Decompiler;
 import org.eolang.opeo.decompilation.DummyDecompiler;
-import org.eolang.opeo.decompilation.LinedDecompiler;
+import org.eolang.opeo.decompilation.FormattingDecompiler;
 import org.eolang.opeo.decompilation.NaiveDecompiler;
 
 /**
@@ -106,7 +106,7 @@ public final class DecompileMojo extends AbstractMojo {
             decompiler = new DummyDecompiler(this.sourcesDir.toPath(), this.outputDir.toPath());
         } else if (Objects.nonNull(this.modifiedDir)) {
             Logger.info(this, "Use selective decompiler");
-            decompiler = new LinedDecompiler(
+            decompiler = new FormattingDecompiler(
                 new SelectiveDecompiler(
                     this.sourcesDir.toPath(), this.outputDir.toPath(), this.modifiedDir.toPath()
                 ),
